Initial Commit
Made maslowmega.h file and for each maslowcnc reference added a maslow_mega_cnc reference that for now does the same thing.  the major hurdle will be changing back from the due timer library that will not be required and making sure it won't be needed.  It is mostly roughed into place, but will need to be tested.